+++ Casa_la_tara.cpp
@@ -13,7 +13,7 @@ Casa_la_tara::Casa_la_tara(): Casa(){
 }
 
 //constructor cu parametri
-Casa_la_tara::Casa_la_tara(const std::string& loc, int nr, int mp, int dim ): Casa(loc, nr, mp){
+Casa_la_tara::Casa_la_tara(const std::string& loc, int nr, float mp, int dim ): Casa(loc, nr, mp){
     if (dim < 0)
         throw CasaTaraException("Dimensiunea gradinii e negativa!");
 
@@ -21,8 +21,8 @@ Casa_la_tara::Casa_la_tara(const std::string& loc, int nr, int mp, int dim ): Ca
 }
 
 //getPret
-int Casa_la_tara::getPret()const {
-    int pret_baza = metri_patrati * Agentie::pret_mp_tara;
+float Casa_la_tara::getPret()const {
+    float pret_baza = metri_patrati * Agentie::pret_mp_tara;
     return pret_baza + pret_baza*taxa_teren;
 }
